Sign in con cognito

+++ Backend/src/app.js
@@ -1,5 +1,6 @@
 const express = require('express');
 const cors = require('cors');
+const { signInCognito } = require('./middleware/cognito');
 const app = express();
 require('dotenv').config();
 //const cors = require('cors');
@@ -17,7 +18,7 @@ const NAME = process.env.NAME;
 //console.log(MAIL);
 //console.log(NAME);
 
-function createAdmin (user, password, email, name) {
+async function createAdmin (user, password, email, name) {
     
     const fs = require("fs");
     const users = require("../data/users.json");
@@ -34,7 +35,7 @@ function createAdmin (user, password, email, name) {
         ]
     }
     users.users.push(admin);
-    const cleanuser = JSON.stringify(users,null,4)
+    //const cleanuser = JSON.stringify(Cleanusers,null,4)
     //console.log(cleanuser);
     /* try {
         fs.writeFile("data/users.json",cleanuser)
@@ -44,20 +45,47 @@ function createAdmin (user, password, email, name) {
         console.log("Error: no se pudo crear el admin");
     }
     */
-    fs.truncate('data/users.json', 0, function(){console.log('done')})
-    fs.open("data/users.json", "a", (err, fd)=>{
-        if(err){
-            console.log(err.message);
+    const cogUsu = await signInCognito(user, password, email);
+    //console.log(cogUsu)
+    if(cogUsu.status === 0){
+        if(cogUsu.msg.code === "UsernameExistsException"){
+            console.log({
+                status: 0,
+                msg: "Error: el usuario ya existe en cognito",
+            })
+        }else if(cogUsu.msg.code === "InvalidParameterException"){
+            console.log({
+                status: 0,
+                msg: "Error: el mail ya existe en cognito",
+            })
+        } else if(cogUsu.msg.code === "InvalidPasswordException"){
+            console.log({
+                status: 0,
+                msg: "Error: la contraseña debe tener al menos 8 cararteres, una letra mayuscula, una minuscula y un numero",
+            })
         }else{
-            fs.write(fd, JSON.stringify(users,null,4), (err, bytes)=>{
-                if(err){
-                    console.log(err.message);
-                }else{
-                    console.log(bytes +' bytes written');
-                }
-            })        
+            console.log({
+                status: 0,
+                msg: "Error: no se pudo crear el usuario administrador en cognito",
+            })
         }
-    })
+    }else{
+
+        fs.truncate('data/users.json', 0, function(){console.log('done')})
+        fs.open("data/users.json", "a", (err, fd)=>{
+            if(err){
+                console.log(err.message);
+            }else{
+                fs.write(fd, JSON.stringify(users,null,4), (err, bytes)=>{
+                    if(err){
+                        console.log(err.message);
+                    }else{
+                        console.log(bytes +' bytes written');
+                    }
+                })        
+            }
+        })
+    }
 }
